backend/requirements.txt
# FastAPI Application Requirements
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pyngrok==7.2.2
python-dotenv==1.0.1

# Existing dependencies (should already be installed)
mistralai>=1.0.0
anthropic>=0.39.0